backend/requirements.txt
fastapi
uvicorn
httpx
python-jose
python-dotenv
